% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-funs.R
\name{as_bam}
\alias{as_bam}
\title{convert \code{GAlignments} object to \code{data.table}}
\usage{
as_bam(x)
}
\arguments{
\item{x}{A \code{GAlignments} object.}
}
\value{
An object of class \code{bam} that inherits from \code{data.table}.
}
\description{
For internal use only. Converts a \code{GAlignments} object to 
\code{bam} object.
}
\examples{
\dontrun{
For internal use only
library(GenomicAlignments)
path <- system.file("tests", package="gread")
bam_file <- file.path(path, "sample.bam")

# with no metadata
bam <- GenomicAlignments::readGAlignments(bam_file)
gread:::as_bam(bam)

# with some metadata
bam <- as(as_granges(read_format(bam_file)), "GAlignments")
gread:::as_bam(bam)
}
}

