% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-features.R
\name{as_data_table}
\alias{as_data_table}
\title{Convert gtf/gff/bam/bed S4 objects to data.table and preserve class}
\usage{
as_data_table(x, reset_class = FALSE)
}
\arguments{
\item{x}{An object that inherits from \code{GRanges}.}

\item{reset_class}{logical (default \code{FALSE}). If \code{TRUE}, resets 
the class to \code{"data.table", "data.frame"}.}
}
\value{
A data.table object with input class preserved.
}
\description{
\code{as.data.table} converts the input object to 
\code{data.table}, but strips away all other classes except 
\code{data.table} and \code{data.frame}. 

This internal function is a simple wrapper to \code{as.data.table} but 
also retains the extra class information of the input object.
}
\examples{
\dontrun{
setClass("foo", contains="GRanges")
x = new("foo", GRanges(letters[1:5], IRanges(1:5, 6:10)))
class(x) # [1] "foo"
class(gread:::as_data_table(x)) # [1] "foo" "data.table" "data.frame"
}
}

