% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_granges.R
\name{as_granges}
\alias{as_granges}
\title{Convert to GRanges object}
\usage{
as_granges(x, ignore_strand = FALSE)
}
\arguments{
\item{x}{An object of class \code{gtf}, \code{gff}, \code{bed} or 
\code{bam} or a valid \code{data.table} object.}

\item{ignore_strand}{Logical argument to pass to \code{GRanges} function. 
Indicates whether \code{strand} should be ignored when constructing 
\code{GRanges} object or not. Default is \code{FALSE}.}
}
\value{
A \code{GRanges} object.
}
\description{
Convert \code{gtf}, \code{gff}, \code{bed}, \code{bam} 
or a valid \code{data.table} to a GRanges object, preserving all 
additional columns.
}
\examples{
path <- system.file("tests", package="gread")
gff_file <- file.path(path, "sample.gff")
gtf_file <- file.path(path, "sample.gtf")
bed_file <- file.path(path, "sample.bed")
bam_file <- file.path(path, "sample.bam")

gff <- read_format(gff_file)
gtf <- read_format(gtf_file)
bed <- read_format(bed_file)
bam <- read_format(bam_file)

as_granges(gff)
as_granges(gtf)
as_granges(bed)
as_granges(bam)

as_granges(gff, ignore_strand=FALSE)
as_granges(gtf, ignore_strand=FALSE)
as_granges(bed, ignore_strand=FALSE)
as_granges(bam, ignore_strand=FALSE)
}
\seealso{
\code{\link{read_format}} 
\code{\link{tidy_cols}} \code{\link{extract}} \code{\link{construct_introns}}
}

