% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-setup.R
\docType{class}
\name{bam-class}
\alias{bam-class}
\title{bam S4 class designed to store data of Binary sequence Alignment/Map Format
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}}
\description{
bam S4 class designed to store data of Binary sequence Alignment/Map Format
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}
}

