% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-setup.R
\docType{class}
\name{bed-class}
\alias{bed-class}
\title{bed S4 class designed to store data of Browser Extensible Data
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}}
\description{
bed S4 class designed to store data of Browser Extensible Data
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}
}

