% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_introns.R
\name{construct_introns}
\alias{construct_introns}
\title{Construct introns from \code{gtf} or \code{gff} objects}
\usage{
construct_introns(x, update = TRUE)
}
\arguments{
\item{x}{An object of class \code{gtf/gff}. It has to have \code{exon} 
feature present.}

\item{update}{If \code{TRUE} (default), \code{x} is updated by reference 
and returned invisibily, else just the \code{intron} coordinates are 
returned.}
}
\value{
An object of class \code{gtf/gff} with updated \code{intron} 
coordinates or just the intron coordinates depending on \code{update}.
}
\description{
This function generates intronic coordinates by extracting 
all the \code{exons} from a \code{gtf} or \code{gff} object.
}
\examples{
path <- system.file("tests", package="gread")
gtf_file <- file.path(path, "sample.gtf")
gtf <- read_format(gtf_file)
# update gtf with intron coordinats from the exon gaps 
# and return the updated object
ans <- construct_introns(gtf, update=TRUE)[] # default
# same as above, but returns only constructed introns
introns <- construct_introns(gtf, update=FALSE)
}
\seealso{
\code{\link{supported_formats}} \code{\link{read_format}} 
\code{\link{extract}} \code{\link{tidy_cols}} \code{\link{as_granges}}
}

