% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-setup.R
\docType{class}
\name{exon-class}
\alias{exon-class}
\title{exon S4 class designed to store exon data
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}}
\description{
exon S4 class designed to store exon data
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}
}

