% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-funs.R
\name{find_overlaps}
\alias{find_overlaps}
\title{Find overlapping indices of two gtf/gff/bed/bam objects}
\usage{
find_overlaps(x, y, ignore_redundant = FALSE, ignore_strand = FALSE, ...)
}
\arguments{
\item{x, y}{An object of class \code{gtf}, \code{gff}, \code{bed} or 
\code{bam}.}

\item{ignore_redundant}{Should redundant overlaps be ignored?}

\item{ignore_strand}{Logical argument to pass to \code{GRanges} function. 
Indicates whether \code{strand} should be ignored when constructing 
\code{GRanges} object or not. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to 
\code{GenomicRanges::findOverlaps}.}
}
\value{
A \code{data.table} containing overlapping indices.
}
\description{
For internal use only. Function for finding overlaps between 
two objects of class \code{gtf/gff/bed/bam} using 
\code{GenomicRanges::findOverlaps}.
}

