% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-setup.R
\docType{class}
\name{gff-class}
\alias{gff-class}
\title{gff S4 class designed to store data of Gene Feature Format
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}}
\description{
gff S4 class designed to store data of Gene Feature Format
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}
}

