% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-setup.R
\docType{class}
\name{gtf-class}
\alias{gtf-class}
\title{gtf S4 class designed to store data of Gene Transfer Format
and dispatch to corresponding functions in \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}}
\description{
gtf S4 class designed to store data of Gene Transfer Format
and dispatch to corresponding functions in \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}
}

