% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-funs.R
\name{intersect_overlaps}
\alias{intersect_overlaps}
\title{Compute intersecting ranges on a gtf/gff/bed/bam object with itself.}
\usage{
intersect_overlaps(x, by = "gene_id", ignore_strand = FALSE)
}
\arguments{
\item{x}{An object of class \code{gtf} or \code{gff}.}

\item{by}{Character vector of column names in \code{x} to \emph{group by}.}

\item{ignore_strand}{Logical argument to pass to \code{GRanges} function. 
Indicates whether \code{strand} should be ignored when constructing 
\code{GRanges} object or not. Default is \code{FALSE}.}
}
\value{
A \code{data.table} with intersecting ranges.
}
\description{
For internal use only. Computes intersecting ranges on a 
\code{gtf/gff} object by converting the input object to a GRangesList 
object and calling \code{disjoin()} (from the \code{GenomicRanges} 
package) on it. Returns an object of same class as input.
}

