% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-setup.R
\docType{class}
\name{intron-class}
\alias{intron-class}
\title{intron S4 class designed to store intron data
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}}
\description{
intron S4 class designed to store intron data
and dispatch to the corresponding functions in the \code{gread} and \code{gcount} packages
inheriting from \linkS4class{GRanges}
}

