% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlaps.R
\name{non_overlaps}
\alias{non_overlaps}
\title{Generate non-overlapping intron coordinates}
\usage{
non_overlaps(x, transcript_id = "transcript_id", gene_id = "gene_id")
}
\arguments{
\item{x}{An object of class \code{gtf/gff} object which has to have 
features \code{intron} and \code{exon}.}

\item{transcript_id}{Column name in \code{x} corresponding to transcript id.}

\item{gene_id}{Column name in \code{x} corresponding to gene id.}
}
\value{
Returns a list of data.table containing \emph{non overlapping 
exons} and \emph{non overlapping introns}.
}
\description{
This function takes an \code{gtf/gff} object. If the 
features \code{exon} and \code{intron} are present, then, it attempts 
to find a set of exon and intron coordinates that do NOT overlap with each 
other. If the intron coordinates are not present, it constructs the introns 
first using \code{gread::construct_introns} and then attempts to extract 
non-overlapping exonic and intronic coordinates. If \code{exon} feature is 
not present, it returns an error.

Note: This function is not exported at the moment.
}
\examples{
\dontrun{
path <- system.file("tests", package="gread")
gtf_file <- file.path(path, "sample.gtf")
gtf <- read_format(gtf_file)
non_overlaps(gtf)
}
}

