% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-funs.R
\name{reduce_overlaps}
\alias{reduce_overlaps}
\title{Compute reduced ranges on a gtf/gff/bed/bam object.}
\usage{
reduce_overlaps(x, by = "gene_id", ignore_strand = FALSE)
}
\arguments{
\item{x}{An object of class \code{gtf/gff/bed/bam}.}

\item{by}{Character vector of column names in \code{x} to \emph{group by}.}

\item{ignore_strand}{Logical argument to pass to \code{GRanges} function. 
Indicates whether \code{strand} should be ignored when constructing 
\code{GRanges} object or not. Default is \code{FALSE}.}
}
\value{
A \code{data.table} with reduced ranges.
}
\description{
For internal use only. Computes reduced ranges on a 
\code{gtf/gff/bed/bam} object by converting the input object to a 
\code{GRangesList} object and calling \code{reduce()} (from 
\code{GenomicRanges} package) on it. Returns an object of same class 
as input.
}

