% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_granges.R
\name{shallow}
\alias{shallow}
\title{shallow copy a \code{data.table}}
\usage{
shallow(x, cols = names(x))
}
\arguments{
\item{x}{A \code{data.table}.}

\item{cols}{Character vector of column names (from \code{x}).}
}
\value{
A shallow copied \code{data.table}.
}
\description{
Convenience function to shallow copy a \code{data.table} 
(until this function is exported in the \code{data.table} package). For 
internal use only.
}
\examples{
\dontrun{
# For internal use only
library(data.table)
x <- data.table(a=1:2, b=3:4)
setattr(x, 'class', c("tmp", class(x)))

y <- gread:::shallow(x) # only copies column pointers
class(y) # class(x) is retained
}
}

