% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported-formats.R
\name{supported_formats}
\alias{supported_formats}
\title{Currently supported formats}
\usage{
supported_formats()
}
\value{
A character vector listing all currently supported formats.
}
\description{
Maintains a vector of all the formats supported currently.
}
\examples{
supported_formats()
}
\seealso{
\code{\link{read_format}}
}

