% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_format.R
\name{tidy}
\alias{tidy}
\alias{tidy.bam}
\alias{tidy.bed}
\alias{tidy.default}
\alias{tidy.gff}
\alias{tidy.gtf}
\title{Tidy up GTF/GFF/BED/BAM objects}
\usage{
tidy(x, ...)

\method{tidy}{default}(x, ...)

\method{tidy}{gtf}(x, remove_cols = "attributes", verbose = FALSE, ...)

\method{tidy}{gff}(x, remove_cols = "attributes", verbose = FALSE, ...)

\method{tidy}{bed}(x, remove_cols = NULL, verbose = FALSE, ...)

\method{tidy}{bam}(x, remove_cols = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{Input object, of class \code{GTF}, \code{GFF}, \code{BED} or 
\code{BAM}.}

\item{...}{Arguments that are ignored at the moment.}

\item{remove_cols}{A character vector of column names to be removed.}

\item{verbose}{Logical. Default is \code{FALSE}. If \code{TRUE}, helpful 
status messages are printed on to the console.}
}
\value{
A tidied object of class \code{GTF}, \code{GFF}, \code{BED} or 
\code{BAM}, that inherits from \code{data.table}.
}
\description{
Tidy up data depending on the type of input format. 

Note that this operation is performed \emph{by reference}, to avoid 
unnecessary copies, for efficiency. Therefore there is no need to assign 
the result back to a variable.
}
\details{
In case of \code{GTF} and \code{GFF} formats, the 
\code{attributes} column is extracted into separate columns after which 
it is removed (by default). Set \code{remove_cols} to \code{NULL} if you 
would like to retain the column even after tidying up.

In case of \code{BED} and \code{BAM} files, no columns are removed by 
default. However, you can use \code{remove_cols} argument if necessary.

In case of \code{BAM} files, the function \code{GAlignments} from the 
\code{GenomicAlignments} Bioconductor package is used to read in the 
file, with the additional arguments for \code{ScanBamParam()} function 
using \code{NM} and \code{MD} tags by default.
}
\examples{
path = system.file("tests", package="gread")
gff_file = file.path(path, "sample.gff")
gtf_file = file.path(path, "sample.gtf")
bed_file = file.path(path, "sample.bed")
bam_file = file.path(path, "sample.bam")

gtf = read_format(gtf_file, tidy=FALSE)
gff = read_format(gtf_file, tidy=FALSE)
bed = read_format(bed_file, tidy=FALSE)
bam = read_format(bam_file, tidy=FALSE)

tidy(gtf, remove_cols=NULL)[]          # tidy attributes col, but not remove
tidy(gff, remove_cols=NULL)[]          # same as above, but for gff
tidy(gtf, remove_cols="attributes")[]  # tidy *and* remove attributes col
tidy(gff, remove_cols="attributes")[]  # same as above, but for gff

tidy(bed, remove_cols="name")[]        # remove name column
tidy(bam, remove_cols=c("NM", "MD"))[] # remove additional loaded tags 
}
\seealso{
\code{\link{supported_formats}} \code{\link{read_format}} 
\code{\link{extract}} \code{\link{construct_introns}} 
\code{\link{as_granges}}
}

