% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-format.R
\name{read_format}
\alias{read_bam}
\alias{read_bed}
\alias{read_format}
\alias{read_gff}
\alias{read_gtf}
\title{Quick and easy reading of gtf/gff/bed/bam files}
\usage{
read_format(file, format = detect_format(file), filter = FALSE,
  chromosomes = NULL, tags = c("NM", "MD"), verbose = FALSE)

read_gtf(file, filter = FALSE, verbose = FALSE)

read_gff(file, filter = FALSE, verbose = FALSE)

read_bed(file, filter = FALSE, verbose = FALSE)

read_bam(file, filter = FALSE, chromosomes = NULL, tags = c("NM", "MD"),
  verbose = FALSE)
}
\arguments{
\item{file}{Complete path to the input file.}

\item{format}{Type of annotation. Currently allowed values are 
\code{"gtf"}, \code{"gff"}, \code{"bed"} and \code{"bam"}. If no input is 
provided, then it will be guessed from input file's extension.}

\item{filter}{Filter rows where \code{start/stop} are invalid, and where 
\code{start > end}. Default (\code{FALSE}) is to not filter, i.e., retain 
all invalid rows as well.}

\item{chromosomes}{Argument to \code{read_bam}. If specified only reads 
from those chromosomes will be loaded.}

\item{tags}{Additional (optional) tags to load from the bam file. By 
default, the tags \code{"NM"} and \code{"MD"} are loaded.}

\item{verbose}{If \code{TRUE}, sends useful status messages to the console. 
Default is \code{FALSE}.}
}
\value{
An object of class \code{gtf}, \code{gff}, \code{bed} or \code{bam}, 
corresponding to the input file format, that inherits from \code{GRanges}.
}
\description{
\code{read_gtf}, \code{read_gff}, \code{read_bed} and 
\code{read_bam} are specific functions corresponding to respective file 
formats. 

\code{read_format} is a top-level convenience function that detects and 
reads automatically from the extension and is sufficient in most cases.
}
\examples{
path <- system.file("tests", package="gread")
gff_file <- file.path(path, "sample.gff")
gtf_file <- file.path(path, "sample.gtf")
bed_file <- file.path(path, "sample.bed")
bam_file <- file.path(path, "sample.bam")

read_format(gff_file) # read gff
read_gff(gff_file)    # same as above
read_format(gtf_file) # read gtf
read_gtf(gtf_file)    # same as above
read_format(bed_file) # read bed
read_bed(bed_file)    # same as above
read_format(bam_file) # read bam
read_bam(bam_file)    # same as above

gtf_filter_file <- file.path(path, "sample_filter.gtf")
read_format(gtf_filter_file, filter=TRUE) # filter invalid rows
read_gtf(gtf_filter_file, filter=TRUE)    # same as above
}
\seealso{
\code{\link{supported_formats}} 
\code{\link{extract}} \code{\link{construct_introns}}
}
\keyword{file}

